lots of changes, ChangeLogs have more info, here are the highlights:
generated makefiles use /pkg properly (will change to -pkg soon)
solution makefile is generated, this is beta, and was tested with one single solution, so prolly wont work for most.
MonoDoc Resolver: if you hit f1, we attempt to find the docs for the type you are playing with. Doesnt do anything but T: style docs, so it needs lots of love.


git-svn-id: svn+ssh://mono-cvs.ximian.com/svn/monodevelop/trunk@1601 b5faff8f-5bce-0310-b493-a43f95cdeebf